Ajustement de DOA joueur + changement de l'INT en Double pour certain attribut

+++ src/main/java/be/helha/projects/GuerreDesRoyaumes/DAOImpl/ItemDAOImpl.java
@@ -36,8 +36,8 @@ public abstract class ItemDAOImpl implements ItemDAO {
                 Arme arme = (Arme) item;
                 String sqlArme = "INSERT INTO armes (item_id, degats) VALUES (?, ?)";
                 try (PreparedStatement statementArme = connection.prepareStatement(sqlArme)) {
-                    statementArme.setInt(1, arme.getId());
-                    statementArme.setInt(2, arme.getDegats());
+                    statementArme.setDouble(1, arme.getId());
+                    statementArme.setDouble(2, arme.getDegats());
                     statementArme.executeUpdate();
                 }
             }
@@ -134,8 +134,8 @@ public abstract class ItemDAOImpl implements ItemDAO {
                 Arme arme = (Arme) item;
                 String sqlArme = "UPDATE armes SET degats = ? WHERE item_id = ?";
                 try (PreparedStatement statementArme = connection.prepareStatement(sqlArme)) {
-                    statementArme.setInt(1, arme.getDegats());
-                    statementArme.setInt(2, arme.getId());
+                    statementArme.setDouble(1, arme.getDegats());
+                    statementArme.setDouble(2, arme.getId());
                     statementArme.executeUpdate();
                 }
             }
